fix movieId field add default messages into constants
