+++ middlewares/auth.js
@@ -1,11 +1,12 @@
 const jwt = require('jsonwebtoken');
 const UnauthorizedError = require('../errors/UnauthorizedError');
+const { UNAUTHORIZED_ERROR_ON_AUTH_CHECK } = require('../utils/constants');
 
 const { JWT_SECRET } = require('../config');
 
 module.exports = (req, res, next) => {
   if (!req.cookies.jwt) {
-    throw new UnauthorizedError('Нужна авторизация');
+    throw new UnauthorizedError(UNAUTHORIZED_ERROR_ON_AUTH_CHECK);
   }
 
   const token = req.cookies.jwt;
@@ -15,7 +16,7 @@ module.exports = (req, res, next) => {
   try {
     payload = jwt.verify(token, JWT_SECRET);
   } catch (err) {
-    throw new UnauthorizedError('Нужна авторизация');
+    throw new UnauthorizedError(UNAUTHORIZED_ERROR_ON_AUTH_CHECK);
   }
 
   req.user = payload;
