+++ middlewares/errorHandler.js
@@ -1,0 +1,14 @@
+const { DEFAULT_SERVER_ERROR_MESSAGE } = require('../utils/constants');
+
+const errorHandler = (err, req, res, next) => {
+  const { statusCode = 500, message } = err;
+
+  res
+    .status(statusCode)
+    .send({
+      message: statusCode === 500 ? DEFAULT_SERVER_ERROR_MESSAGE : message,
+    });
+  next();
+};
+
+module.exports = errorHandler;
